refactor(common): split util helpers by responsibility
