+++ lib/common/util.go
@@ -2,16 +2,12 @@ package common
 
 import (
 	"bytes"
-	"crypto/rand"
-	"crypto/sha256"
 	"crypto/tls"
 	"encoding/base64"
 	"encoding/binary"
-	"fmt"
 	"html/template"
 	"io"
 	"math"
-	"math/big"
 	"net"
 	"net/http"
 	"net/url"
@@ -23,10 +19,8 @@ import (
 	"strings"
 	"sync"
 	"time"
-	_ "time/tzdata"
 
 	"github.com/araddon/dateparse"
-	"github.com/beevik/ntp"
 	"github.com/djylb/nps/lib/logs"
 )
 
@@ -35,8 +29,6 @@ var (
 	domainCheckRegexp         = regexp.MustCompile(`^((http://)|(https://))?([a-zA-Z0-9]([a-zA-Z0-9\-]{0,61}[a-zA-Z0-9])?\.)+[a-zA-Z]{2,6}`)
 )
 
-// ExtractHost
-// return "[2001:db8::1]:80"
 func ExtractHost(input string) string {
 	if strings.Contains(input, "://") {
 		if u, err := url.Parse(input); err == nil && u.Host != "" {
@@ -49,59 +41,46 @@ func ExtractHost(input string) string {
 	return input
 }
 
-// RemovePortFromHost
-// return "[2001:db8::1]"
 func RemovePortFromHost(host string) string {
 	if len(host) == 0 {
 		return host
 	}
 	var idx int
-	// IPv6
 	if host[0] == '[' {
 		if idx = strings.IndexByte(host, ']'); idx != -1 {
 			return host[:idx+1]
 		}
 		return ""
 	}
-	// IPv4 or Domain
 	if idx = strings.LastIndexByte(host, ':'); idx != -1 && idx == strings.IndexByte(host, ':') {
 		return host[:idx]
 	}
 	return host
 }
 
-// GetIpByAddr
-// return "2001:db8::1"
 func GetIpByAddr(host string) string {
 	if len(host) == 0 {
 		return host
 	}
 	var idx int
-	// IPv6
 	if host[0] == '[' {
 		if idx = strings.IndexByte(host, ']'); idx != -1 {
 			return host[1:idx]
 		}
 		return ""
 	}
-	// IPv4 or Domain
 	if idx = strings.LastIndexByte(host, ':'); idx != -1 && idx == strings.IndexByte(host, ':') {
 		return host[:idx]
 	}
 	return host
 }
 
-func IsDomain(s string) bool {
-	return net.ParseIP(s) == nil
-}
+func IsDomain(s string) bool { return net.ParseIP(s) == nil }
 
-// GetPortByAddr
-// return int or 0
 func GetPortByAddr(addr string) int {
 	if len(addr) == 0 {
 		return 0
 	}
-	// IPv6
 	if addr[0] == '[' {
 		if end := strings.IndexByte(addr, ']'); end != -1 && end+1 < len(addr) && addr[end+1] == ':' {
 			portStr := addr[end+2:]
@@ -111,7 +90,6 @@ func GetPortByAddr(addr string) int {
 		}
 		return 0
 	}
-	// Other
 	if idx := strings.LastIndexByte(addr, ':'); idx != -1 {
 		portStr := addr[idx+1:]
 		if port, err := strconv.Atoi(portStr); err == nil && port > 0 && port <= 65535 {
@@ -145,7 +123,7 @@ func ValidateAddr(s string) string {
 }
 
 func BuildAddress(host string, port string) string {
-	if strings.Contains(host, ":") { // IPv6
+	if strings.Contains(host, ":") {
 		return "[" + host + "]:" + port
 	}
 	return host + ":" + port
@@ -181,7 +159,6 @@ func GetSni(host string) string {
 	return sni
 }
 
-// GetHostByName Get the corresponding IP address through domain name
 func GetHostByName(hostname string) string {
 	if !DomainCheck(hostname) {
 		return hostname
@@ -191,18 +168,13 @@ func GetHostByName(hostname string) string {
 		return ""
 	}
 	for _, v := range ips {
-		if v.To4() != nil {
-			return v.String()
-		}
-		// If IPv4 not found, return IPv6
-		if v.To16() != nil {
+		if v.To4() != nil || v.To16() != nil {
 			return v.String()
 		}
 	}
 	return ""
 }
 
-// DomainCheck Check the legality of domain
 func DomainCheck(domain string) bool {
 	return domainCheckWithPathRegexp.MatchString(domain) || domainCheckRegexp.MatchString(domain)
 }
@@ -234,51 +206,35 @@ func GetPort(value int) int {
 	return (65536 + value%65536) % 65536
 }
 
-// CheckAuthWithAccountMap
-// u current login user
-// p current login passwd
-// user global user
-// passwd global passwd
-// accountMap enable multi user auth
 func CheckAuthWithAccountMap(u, p, user, passwd string, accountMap, authMap map[string]string) bool {
-	// Single account check
 	noAccountMap := len(accountMap) == 0
 	noAuthMap := len(authMap) == 0
 	if noAccountMap && noAuthMap {
 		return u == user && p == passwd
 	}
-
-	// Multi-account authentication check
 	if len(u) == 0 {
 		return false
 	}
-
 	if u == user && p == passwd {
 		return true
 	}
-
 	if !noAccountMap {
 		if P, ok := accountMap[u]; ok && p == P {
 			return true
 		}
 	}
-
 	if !noAuthMap {
 		if P, ok := authMap[u]; ok && p == P {
 			return true
 		}
 	}
-
 	return false
 }
 
-// CheckAuth Check if the Request request is validated
 func CheckAuth(r *http.Request, user, passwd string, accountMap, authMap map[string]string) bool {
-	// Bypass authentication only if user, passwd are empty and multiAccount is nil or empty
 	if user == "" && passwd == "" && len(accountMap) == 0 && len(authMap) == 0 {
 		return true
 	}
-
 	s := strings.SplitN(r.Header.Get("Authorization"), " ", 2)
 	if len(s) != 2 {
 		s = strings.SplitN(r.Header.Get("Proxy-Authorization"), " ", 2)
@@ -286,17 +242,14 @@ func CheckAuth(r *http.Request, user, passwd string, accountMap, authMap map[str
 			return false
 		}
 	}
-
 	b, err := base64.StdEncoding.DecodeString(s[1])
 	if err != nil {
 		return false
 	}
-
 	pair := strings.SplitN(string(b), ":", 2)
 	if len(pair) != 2 {
 		return false
 	}
-
 	return CheckAuthWithAccountMap(pair[0], pair[1], user, passwd, accountMap, authMap)
 }
 
@@ -305,7 +258,6 @@ func DealMultiUser(s string) map[string]string {
 	if s == "" {
 		return nil
 	}
-
 	s = strings.ReplaceAll(s, "\r\n", "\n")
 	multiUserMap := make(map[string]string)
 	for _, v := range strings.Split(s, "\n") {
@@ -323,16 +275,8 @@ func DealMultiUser(s string) map[string]string {
 	return multiUserMap
 }
 
-// GetBoolByStr get bool by str
-func GetBoolByStr(s string) bool {
-	switch s {
-	case "1", "true":
-		return true
-	}
-	return false
-}
+func GetBoolByStr(s string) bool { return s == "1" || s == "true" }
 
-// GetStrByBool get str by bool
 func GetStrByBool(b bool) string {
 	if b {
 		return "1"
@@ -340,37 +284,26 @@ func GetStrByBool(b bool) string {
 	return "0"
 }
 
-// GetIntNoErrByStr int
 func GetIntNoErrByStr(str string) int {
 	i, _ := strconv.Atoi(strings.TrimSpace(str))
 	return i
 }
 
-// GetTimeNoErrByStr time
 func GetTimeNoErrByStr(str string) time.Time {
-	// 1. 去除前后空格
 	str = strings.TrimSpace(str)
 	if str == "" {
-		return time.Time{} // 为空时返回零时间
+		return time.Time{}
 	}
-
-	// 2. 先尝试解析为 Unix 时间戳（秒或毫秒）
 	if timestamp, err := strconv.ParseInt(str, 10, 64); err == nil {
-		// 处理毫秒级时间戳
 		if timestamp > 1_000_000_000_000 {
 			return time.UnixMilli(timestamp)
 		}
-		// 处理秒级时间戳
 		return time.Unix(timestamp, 0)
 	}
-
-	// 3. 使用 dateparse 库解析日期字符串
 	t, err := dateparse.ParseLocal(str)
 	if err == nil {
 		return t
 	}
-
-	// 解析失败，返回零时间
 	return time.Time{}
 }
 
@@ -378,7 +311,6 @@ func ContainsFold(s, substr string) bool {
 	return strings.Contains(strings.ToLower(s), strings.ToLower(substr))
 }
 
-// ReadAllFromFile Read file content by file path
 func ReadAllFromFile(filePath string) ([]byte, error) {
 	f, err := os.Open(filePath)
 	if err != nil {
@@ -416,7 +348,6 @@ func GetCertContent(filePath, header string) (string, error) {
 func LoadCertPair(certFile, keyFile string) (certContent, keyContent string, ok bool) {
 	var wg sync.WaitGroup
 	var certErr, keyErr error
-
 	wg.Add(2)
 	go func() {
 		defer wg.Done()
@@ -427,7 +358,6 @@ func LoadCertPair(certFile, keyFile string) (certContent, keyContent string, ok 
 		keyContent, keyErr = GetCertContent(keyFile, "PRIVATE")
 	}()
 	wg.Wait()
-
 	if certErr != nil || keyErr != nil || certContent == "" || keyContent == "" {
 		return "", "", false
 	}
@@ -458,17 +388,13 @@ func GetCertType(s string) string {
 	return "invalid"
 }
 
-// FileExists reports whether the named file or directory exists.
 func FileExists(name string) bool {
-	if _, err := os.Stat(name); err != nil {
-		if os.IsNotExist(err) {
-			return false
-		}
+	if _, err := os.Stat(name); err != nil && os.IsNotExist(err) {
+		return false
 	}
 	return true
 }
 
-// TestTcpPort Judge whether the TCP port can open normally
 func TestTcpPort(port int) bool {
 	l, err := net.ListenTCP("tcp", &net.TCPAddr{IP: net.ParseIP("0.0.0.0"), Port: port})
 	defer func() {
@@ -479,7 +405,6 @@ func TestTcpPort(port int) bool {
 	return err == nil
 }
 
-// TestUdpPort Judge whether the UDP port can open normally
 func TestUdpPort(port int) bool {
 	l, err := net.ListenUDP("udp", &net.UDPAddr{IP: net.ParseIP("0.0.0.0"), Port: port})
 	defer func() {
@@ -490,9 +415,6 @@ func TestUdpPort(port int) bool {
 	return err == nil
 }
 
-// BinaryWrite Write length and individual byte data
-// Length prevents sticking
-// # Characters are used to separate data
 func BinaryWrite(raw *bytes.Buffer, v ...string) {
 	b := GetWriteStr(v...)
 	var lenBuf [4]byte
@@ -501,7 +423,6 @@ func BinaryWrite(raw *bytes.Buffer, v ...string) {
 	_, _ = raw.Write(b)
 }
 
-// GetWriteStr get seq str
 func GetWriteStr(v ...string) []byte {
 	sep := CONN_DATA_SEQ
 	sepLen := len(sep)
@@ -509,7 +430,6 @@ func GetWriteStr(v ...string) []byte {
 	for _, s := range v {
 		total += len(s) + sepLen
 	}
-
 	buffer := make([]byte, 0, total)
 	for _, s := range v {
 		buffer = append(buffer, s...)
@@ -518,7 +438,6 @@ func GetWriteStr(v ...string) []byte {
 	return buffer
 }
 
-// InStrArr inArray str interface
 func InStrArr(arr []string, val string) bool {
 	for _, v := range arr {
 		if v == val {
@@ -528,7 +447,6 @@ func InStrArr(arr []string, val string) bool {
 	return false
 }
 
-// InIntArr inArray int interface
 func InIntArr(arr []int, val int) bool {
 	for _, v := range arr {
 		if v == val {
@@ -538,7 +456,6 @@ func InIntArr(arr []int, val int) bool {
 	return false
 }
 
-// GetPorts format ports str to an int array
 func GetPorts(s string) []int {
 	if strings.TrimSpace(s) == "" {
 		return nil
@@ -579,19 +496,14 @@ func GetPorts(s string) []int {
 	return ps
 }
 
-// IsPort is the string a port
 func IsPort(p string) bool {
 	pi, err := strconv.Atoi(p)
 	if err != nil {
 		return false
 	}
-	if pi > 65536 || pi < 1 {
-		return false
-	}
-	return true
+	return pi <= 65536 && pi >= 1
 }
 
-// FormatAddress if the s is just a port,return 127.0.0.1:s
 func FormatAddress(s string) string {
 	if strings.Contains(s, ":") {
 		return s
@@ -602,10 +514,7 @@ func FormatAddress(s string) string {
 func in(target string, strArray []string) bool {
 	sort.Strings(strArray)
 	index := sort.SearchStrings(strArray, target)
-	if index < len(strArray) && strArray[index] == target {
-		return true
-	}
-	return false
+	return index < len(strArray) && strArray[index] == target
 }
 
 func IsBlackIp(ipPort, vkey string, blackIpList []string) bool {
@@ -622,9 +531,6 @@ func CopyBuffer(dst io.Writer, src io.Reader, label ...string) (written int64, e
 	defer BufPoolCopy.Put(buf)
 	for {
 		nr, er := src.Read(buf)
-		//if len(pr)>0 && pr[0] && nr > 50 {
-		//	logs.Warn(string(buf[:50]))
-		//}
 		if nr > 0 {
 			nw, ew := dst.Write(buf[0:nr])
 			if nw > 0 {
@@ -647,7 +553,6 @@ func CopyBuffer(dst io.Writer, src io.Reader, label ...string) (written int64, e
 	return written, err
 }
 
-// GetLocalUdpAddr send this ip forget to get a local udp port
 func GetLocalUdpAddr() (net.Conn, error) {
 	tmpConn, err := net.Dial("udp", GetCustomDNS())
 	if err != nil {
@@ -672,40 +577,11 @@ func GetLocalUdp6Addr() (net.Conn, error) {
 	return tmpConn, tmpConn.Close()
 }
 
-// GetLocalUdp4IP returns the preferred local IPv4 egress IP.
-func GetLocalUdp4IP() (net.IP, error) {
-	c, err := net.Dial("udp4", IPv4DNS)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = c.Close() }()
-	la, ok := c.LocalAddr().(*net.UDPAddr)
-	if !ok || la == nil || la.IP == nil {
-		return nil, fmt.Errorf("get local udp4 ip failed")
-	}
-	return la.IP, nil
-}
-
-// GetLocalUdp6IP returns the preferred local IPv6 egress IP.
-func GetLocalUdp6IP() (net.IP, error) {
-	c, err := net.Dial("udp6", IPv6DNS)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = c.Close() }()
-	la, ok := c.LocalAddr().(*net.UDPAddr)
-	if !ok || la == nil || la.IP == nil {
-		return nil, fmt.Errorf("get local udp6 ip failed")
-	}
-	return la.IP, nil
-}
-
-// ParseStr parse template
 func ParseStr(str string) (string, error) {
 	tmp := template.New("npc")
-	var err error
 	w := new(bytes.Buffer)
-	if tmp, err = tmp.Parse(str); err != nil {
+	tmp, err := tmp.Parse(str)
+	if err != nil {
 		return "", err
 	}
 	if err = tmp.Execute(w, GetEnvMap()); err != nil {
@@ -714,12 +590,10 @@ func ParseStr(str string) (string, error) {
 	return w.String(), nil
 }
 
-// GetEnvMap get env
 func GetEnvMap() map[string]string {
 	m := make(map[string]string)
-	environ := os.Environ()
-	for i := range environ {
-		tmp := strings.Split(environ[i], "=")
+	for _, kv := range os.Environ() {
+		tmp := strings.Split(kv, "=")
 		if len(tmp) == 2 {
 			m[tmp[0]] = tmp[1]
 		}
@@ -727,11 +601,10 @@ func GetEnvMap() map[string]string {
 	return m
 }
 
-// TrimArr throw the empty element of the string array
 func TrimArr(arr []string) []string {
 	newArr := make([]string, 0)
 	for _, v := range arr {
-		trimmed := strings.TrimSpace(v) // 去除前后空白
+		trimmed := strings.TrimSpace(v)
 		if trimmed != "" {
 			newArr = append(newArr, trimmed)
 		}
@@ -751,12 +624,10 @@ func IsArrContains(arr []string, val string) bool {
 	return false
 }
 
-// RemoveArrVal remove value from string array
 func RemoveArrVal(arr []string, val string) []string {
 	for k, v := range arr {
 		if v == val {
-			arr = append(arr[:k], arr[k+1:]...)
-			return arr
+			return append(arr[:k], arr[k+1:]...)
 		}
 	}
 	return arr
@@ -766,14 +637,12 @@ func HandleArrEmptyVal(list []string) []string {
 	for len(list) > 0 && (list[len(list)-1] == "" || strings.TrimSpace(list[len(list)-1]) == "") {
 		list = list[:len(list)-1]
 	}
-
 	for i := 0; i < len(list); i++ {
 		list[i] = strings.TrimSpace(list[i])
 		if i > 0 && list[i] == "" {
 			list[i] = list[i-1]
 		}
 	}
-
 	return list
 }
 
@@ -784,11 +653,9 @@ func ExtendArrs(arrays ...*[]string) int {
 			maxLength = len(*arr)
 		}
 	}
-
 	if maxLength == 0 {
 		return 0
 	}
-
 	for _, arr := range arrays {
 		for len(*arr) < maxLength {
 			if len(*arr) == 0 {
@@ -798,752 +665,5 @@ func ExtendArrs(arrays ...*[]string) int {
 			}
 		}
 	}
-
 	return maxLength
-}
-
-// BytesToNum convert bytes to num
-func BytesToNum(b []byte) int {
-	var str string
-	for i := 0; i < len(b); i++ {
-		str += strconv.Itoa(int(b[i]))
-	}
-	x, _ := strconv.Atoi(str)
-	return x
-}
-
-// GetSyncMapLen get the length of the sync map
-func GetSyncMapLen(m *sync.Map) int {
-	var c int
-	m.Range(func(key, value interface{}) bool {
-		c++
-		return true
-	})
-	return c
-}
-
-func GetExtFromPath(path string) string {
-	s := strings.Split(path, ".")
-	re, err := regexp.Compile(`(\w+)`)
-	if err != nil {
-		return ""
-	}
-	return string(re.Find([]byte(s[0])))
-}
-
-func NormalizeIP(ip net.IP) net.IP {
-	if ip == nil {
-		return nil
-	}
-	if v4 := ip.To4(); v4 != nil {
-		return v4
-	}
-	return ip.To16()
-}
-
-func IsZeroIP(ip net.IP) bool {
-	if ip == nil {
-		return true
-	}
-	return ip.Equal(net.IPv4zero) || ip.Equal(net.IPv6zero)
-}
-
-func BuildUdpBindAddr(serverIP string, clientIP net.IP) (network string, addr *net.UDPAddr) {
-	if ip := net.ParseIP(serverIP); ip != nil && !IsZeroIP(ip) {
-		if ip.To4() != nil {
-			return "udp4", &net.UDPAddr{IP: ip, Port: 0}
-		}
-		return "udp6", &net.UDPAddr{IP: ip, Port: 0}
-	}
-	if clientIP != nil {
-		if clientIP.To4() != nil {
-			return "udp4", &net.UDPAddr{IP: net.IPv4zero, Port: 0}
-		}
-		return "udp6", &net.UDPAddr{IP: net.IPv6unspecified, Port: 0}
-	}
-	return "udp", &net.UDPAddr{IP: nil, Port: 0}
-}
-
-func IsSameIPType(addr1, addr2 string) bool {
-	ip1 := strings.Contains(addr1, "[")
-	ip2 := strings.Contains(addr2, "[")
-
-	return ip1 == ip2
-}
-
-func GetMatchingLocalAddr(remoteAddr, localAddr string) (string, error) {
-	remoteIsV6 := strings.Contains(remoteAddr, "]:")
-	localIsV6 := strings.Contains(localAddr, "]:")
-	if remoteIsV6 == localIsV6 {
-		return localAddr, nil
-	}
-	port := GetPortStrByAddr(localAddr)
-	if remoteIsV6 {
-		tmpConn, err := GetLocalUdp6Addr()
-		if err != nil {
-			return localAddr, fmt.Errorf("get local ipv6 addr: %w", err)
-		}
-		ip6 := tmpConn.LocalAddr().(*net.UDPAddr).IP.String()
-		return "[" + ip6 + "]:" + port, nil
-	} else {
-		tmpConn, err := GetLocalUdp4Addr()
-		if err != nil {
-			return localAddr, fmt.Errorf("get local ipv4 addr: %w", err)
-		}
-		ip4 := tmpConn.LocalAddr().(*net.UDPAddr).IP.String()
-		return ip4 + ":" + port, nil
-	}
-}
-
-var externalIp string
-var ipApis = []string{
-	"https://4.ipw.cn",
-	"https://api.ipify.org",
-	"http://ipinfo.io/ip",
-	"https://api64.ipify.org",
-	"https://6.ipw.cn",
-	"http://api.ip.sb",
-	"http://myexternalip.com/raw",
-	"http://ifconfig.me/ip",
-	"http://ident.me",
-	"https://d-jy.net/ip",
-}
-
-func FetchExternalIp() string {
-	for _, api := range ipApis {
-		resp, err := http.Get(api)
-		if err != nil {
-			continue
-		}
-		content, _ := io.ReadAll(resp.Body)
-		_ = resp.Body.Close()
-		ip := string(content)
-		if IsValidIP(ip) {
-			externalIp = ip
-			return ip
-		}
-	}
-
-	return ""
-}
-
-func GetExternalIp() string {
-	if externalIp != "" {
-		return externalIp
-	}
-	return FetchExternalIp()
-}
-
-func PickEgressIPFor(dstIP net.IP) net.IP {
-	if dstIP == nil {
-		return nil
-	}
-	network := "udp4"
-	if dstIP.To4() == nil {
-		network = "udp6"
-	}
-	raddr := (&net.UDPAddr{IP: dstIP, Port: 9}).String()
-	d := net.Dialer{Timeout: 300 * time.Millisecond}
-	conn, err := d.Dial(network, raddr)
-	if err != nil {
-		return nil
-	}
-	defer func() { _ = conn.Close() }()
-	if la, ok := conn.LocalAddr().(*net.UDPAddr); ok && la != nil && !IsZeroIP(la.IP) {
-		return la.IP
-	}
-	return nil
-}
-
-func GetIntranetIp() string {
-	addrs, err := net.InterfaceAddrs()
-	if err != nil {
-		return "127.0.0.1"
-	}
-	for _, address := range addrs {
-		// 检查 IP 地址判断是否为回环地址
-		if ipnet, ok := address.(*net.IPNet); ok && !ipnet.IP.IsLoopback() {
-			if ipnet.IP.To4() != nil || ipnet.IP.To16() != nil {
-				return ipnet.IP.String()
-			}
-		}
-	}
-	return "127.0.0.1"
-}
-
-func GetOutboundIP() net.IP {
-	conn, err := net.Dial("udp", GetCustomDNS())
-	if err != nil {
-		return net.ParseIP("127.0.0.1")
-	}
-	defer func() { _ = conn.Close() }()
-	localAddr := conn.LocalAddr().(*net.UDPAddr)
-	return localAddr.IP
-}
-
-func GetOutboundIPv6() net.IP {
-	tmpConn, err := GetLocalUdp6Addr()
-	if err == nil {
-		return tmpConn.LocalAddr().(*net.UDPAddr).IP
-	}
-	return nil
-}
-
-func IsValidIP(ip string) bool {
-	parsedIP := net.ParseIP(ip)
-	return parsedIP != nil
-}
-
-func IsPublicIP(ip net.IP) bool {
-	if ip.IsLoopback() || ip.IsLinkLocalMulticast() || ip.IsLinkLocalUnicast() {
-		return false
-	}
-
-	if ip4 := ip.To4(); ip4 != nil {
-		switch {
-		case ip4[0] == 10:
-			return false
-		case ip4[0] == 172 && ip4[1] >= 16 && ip4[1] <= 31:
-			return false
-		case ip4[0] == 192 && ip4[1] == 168:
-			return false
-		default:
-			return true
-		}
-	}
-
-	// IPv6
-	if ip6 := ip.To16(); ip6 != nil {
-		return !ip6.IsPrivate()
-	}
-
-	return false
-}
-
-func GetServerIp(ip string) string {
-	if ip != "" && ip != "0.0.0.0" && ip != "::" {
-		return ip
-	}
-
-	if ip == "::" {
-		tmpConn, err := GetLocalUdp6Addr()
-		if err == nil {
-			return tmpConn.LocalAddr().(*net.UDPAddr).IP.String()
-		}
-	}
-
-	return GetOutboundIP().String()
-}
-
-func GetServerIpByClientIp(clientIp net.IP) string {
-	if IsPublicIP(clientIp) {
-		return GetExternalIp()
-	}
-	return GetIntranetIp()
-}
-
-// EncodeIP encodes a net.IP to [1-byte ATYP] + [16-byte Address]
-func EncodeIP(ip net.IP) []byte {
-	buf := make([]byte, 17)
-	if ip4 := ip.To4(); ip4 != nil {
-		buf[0] = 0x01
-		copy(buf[1:], ip4)
-	} else {
-		buf[0] = 0x04
-		copy(buf[1:], ip.To16())
-	}
-	return buf
-}
-
-// DecodeIP decodes a [1-byte ATYP] + [16-byte Address] to net.IP
-func DecodeIP(data []byte) net.IP {
-	if len(data) < 17 {
-		return nil
-	}
-	atyp := data[0]
-	addr := data[1:17]
-	switch atyp {
-	case 0x01:
-		return net.IPv4(addr[0], addr[1], addr[2], addr[3])
-	case 0x04:
-		return addr
-	default:
-		return nil
-	}
-}
-
-func JoinHostPort(host string, port string) string {
-	return net.JoinHostPort(host, port)
-}
-
-func RandomBytes(maxLen int) ([]byte, error) {
-	nBig, err := rand.Int(rand.Reader, big.NewInt(int64(maxLen+1)))
-	if err != nil {
-		return nil, err
-	}
-	n := int(nBig.Int64())
-	buf := make([]byte, n)
-	if _, err := rand.Read(buf); err != nil {
-		return nil, err
-	}
-	return buf, nil
-}
-
-func SetTimezone(tz string) error {
-	if tz == "" {
-		return nil
-	}
-	loc, err := time.LoadLocation(tz)
-	if err != nil {
-		return err
-	}
-	time.Local = loc
-	return nil
-}
-
-var (
-	timeOffset   time.Duration
-	ntpServer    string
-	syncInterval = 5 * time.Minute
-	lastSyncMono time.Time
-	timeMutex    sync.RWMutex
-	syncCh       = make(chan struct{}, 1)
-)
-
-func SetNtpServer(server string) {
-	timeMutex.Lock()
-	defer timeMutex.Unlock()
-	ntpServer = server
-}
-
-func SetNtpInterval(d time.Duration) {
-	timeMutex.Lock()
-	defer timeMutex.Unlock()
-	syncInterval = d
-}
-
-func CalibrateTimeOffset(server string) (time.Duration, error) {
-	if server == "" {
-		return 0, nil
-	}
-	ntpTime, err := ntp.Time(server)
-	if err != nil {
-		return 0, err
-	}
-	return time.Until(ntpTime), nil
-}
-
-func TimeOffset() time.Duration {
-	timeMutex.RLock()
-	defer timeMutex.RUnlock()
-	return timeOffset
-}
-
-func TimeNow() time.Time {
-	SyncTime()
-	timeMutex.RLock()
-	defer timeMutex.RUnlock()
-	return time.Now().Add(timeOffset)
-}
-
-func SyncTime() {
-	timeMutex.RLock()
-	srv, last := ntpServer, lastSyncMono
-	interval := syncInterval
-	timeMutex.RUnlock()
-
-	if srv == "" || (!last.IsZero() && time.Since(last) < interval) {
-		return
-	}
-
-	select {
-	case syncCh <- struct{}{}:
-		defer func() { <-syncCh }()
-	default:
-		return
-	}
-
-	now := time.Now()
-	timeMutex.Lock()
-	lastSyncMono = now
-	timeMutex.Unlock()
-
-	offset, err := CalibrateTimeOffset(srv)
-	if err != nil {
-		logs.Error("ntp[%s] sync failed: %v", srv, err)
-	}
-
-	timeMutex.Lock()
-	timeOffset = offset
-	timeMutex.Unlock()
-
-	if offset != 0 {
-		logs.Info("ntp[%s] offset=%v", srv, offset)
-	}
-}
-
-// TimestampToBytes 8bit
-func TimestampToBytes(ts int64) []byte {
-	b := make([]byte, 8)
-	binary.BigEndian.PutUint64(b, uint64(ts))
-	return b
-}
-
-// BytesToTimestamp 8bit
-func BytesToTimestamp(b []byte) int64 {
-	return int64(binary.BigEndian.Uint64(b))
-}
-
-func ValidatePoW(bits int, parts ...string) bool {
-	if bits < 1 || bits > 256 {
-		return false
-	}
-
-	data := strings.Join(parts, "")
-	sum := sha256.Sum256([]byte(data))
-	fullBytes := bits / 8
-	for i := 0; i < fullBytes; i++ {
-		if sum[i] != 0 {
-			return false
-		}
-	}
-	remBits := bits % 8
-	if remBits > 0 {
-		mask := byte(0xFF << (8 - remBits))
-		if (sum[fullBytes] & mask) != 0 {
-			return false
-		}
-	}
-	return true
-}
-
-func IsTrustedProxy(list, ipStr string) bool {
-	if list == "" || ipStr == "" {
-		return false
-	}
-
-	ipStr = strings.TrimSpace(ipStr)
-
-	if h, _, err := net.SplitHostPort(ipStr); err == nil {
-		ipStr = h
-	}
-	if strings.HasPrefix(ipStr, "[") && strings.HasSuffix(ipStr, "]") {
-		ipStr = ipStr[1 : len(ipStr)-1]
-	}
-	if i := strings.LastIndex(ipStr, "%"); i != -1 { // fe80::1%eth0
-		ipStr = ipStr[:i]
-	}
-
-	ip := net.ParseIP(ipStr)
-	if ip == nil {
-		return false
-	}
-	ip4 := ip.To4()
-
-	for _, raw := range strings.Split(list, ",") {
-		entry := strings.TrimSpace(raw)
-		if entry == "" {
-			continue
-		}
-
-		if entry == "*" {
-			return true
-		}
-
-		// CIDR（IPv4/IPv6）
-		if strings.Contains(entry, "/") {
-			if _, cidr, err := net.ParseCIDR(entry); err == nil && cidr.Contains(ip) {
-				return true
-			}
-			continue
-		}
-
-		// if "192.168.*.*"
-		if strings.Contains(entry, "*") {
-			if ip4 == nil {
-				continue
-			}
-			pSegs := strings.Split(entry, ".")
-			if len(pSegs) != 4 {
-				continue
-			}
-			matched := true
-			for i := 0; i < 4; i++ {
-				if pSegs[i] == "*" {
-					continue
-				}
-				n, err := strconv.Atoi(pSegs[i])
-				if err != nil || n < 0 || n > 255 || int(ip4[i]) != n {
-					matched = false
-					break
-				}
-			}
-			if matched {
-				return true
-			}
-			continue
-		}
-
-		if e := net.ParseIP(entry); e != nil && e.Equal(ip) {
-			return true
-		}
-	}
-
-	return false
-}
-
-// SplitCommaAddrList splits "ip:port,ip:port,..." into a de-duplicated list of valid IP:port.
-// - Accepts single item without comma.
-// - Ignores invalid entries.
-// - Keeps first valid one as stable fallback.
-func SplitCommaAddrList(s string) []string {
-	if s == "" {
-		return nil
-	}
-	parts := strings.Split(s, ",")
-	out := make([]string, 0, len(parts))
-	seen := make(map[string]struct{}, len(parts))
-	for _, p := range parts {
-		p = strings.TrimSpace(p)
-		if p == "" {
-			continue
-		}
-		v := ValidateAddr(p)
-		if v == "" {
-			continue
-		}
-		if _, ok := seen[v]; ok {
-			continue
-		}
-		seen[v] = struct{}{}
-		out = append(out, v)
-	}
-	return out
-}
-
-// ParseIPFromAddr parses IP from an "ip:port" (supports "[v6]:port").
-// Returns nil if cannot parse.
-func ParseIPFromAddr(addr string) net.IP {
-	if addr == "" {
-		return nil
-	}
-	ipStr := GetIpByAddr(addr)
-	if ipStr == "" {
-		return nil
-	}
-	if strings.HasPrefix(ipStr, "[") && strings.HasSuffix(ipStr, "]") {
-		ipStr = ipStr[1 : len(ipStr)-1]
-	}
-	if i := strings.LastIndex(ipStr, "%"); i != -1 { // fe80::1%eth0
-		ipStr = ipStr[:i]
-	}
-	return net.ParseIP(ipStr)
-}
-
-// IsPublicIPStrict checks if IP is globally routable on the Internet.
-// It excludes RFC1918, loopback, link-local, multicast, unspecified,
-// and also excludes CGNAT / TEST-NET / benchmarking / documentation ranges.
-func IsPublicIPStrict(ip net.IP) bool {
-	if ip == nil {
-		return false
-	}
-	ip = NormalizeIP(ip)
-	if ip == nil {
-		return false
-	}
-	if !ip.IsGlobalUnicast() {
-		return false
-	}
-	if ip.IsPrivate() || ip.IsLoopback() || ip.IsLinkLocalUnicast() || ip.IsLinkLocalMulticast() || ip.IsMulticast() || ip.IsUnspecified() {
-		return false
-	}
-
-	if ip4 := ip.To4(); ip4 != nil {
-		// CGNAT 100.64.0.0/10
-		if ip4[0] == 100 && ip4[1] >= 64 && ip4[1] <= 127 {
-			return false
-		}
-		// TEST-NET-1/2/3: 192.0.2.0/24, 198.51.100.0/24, 203.0.113.0/24
-		if ip4[0] == 192 && ip4[1] == 0 && ip4[2] == 2 {
-			return false
-		}
-		if ip4[0] == 198 && ip4[1] == 51 && ip4[2] == 100 {
-			return false
-		}
-		if ip4[0] == 203 && ip4[1] == 0 && ip4[2] == 113 {
-			return false
-		}
-		// 198.18.0.0/15 benchmark
-		if ip4[0] == 198 && (ip4[1] == 18 || ip4[1] == 19) {
-			return false
-		}
-		// 192.88.99.0/24 (6to4 relay anycast, deprecated)
-		if ip4[0] == 192 && ip4[1] == 88 && ip4[2] == 99 {
-			return false
-		}
-		return true
-	}
-
-	// IPv6 documentation: 2001:db8::/32
-	if len(ip) >= 4 && ip[0] == 0x20 && ip[1] == 0x01 && ip[2] == 0x0d && ip[3] == 0xb8 {
-		return false
-	}
-	return true
-}
-
-// PickBestV4V6FromLocalList picks best IPv4/IPv6 address from localStr (comma-separated).
-// Rules within each family:
-// - First seen becomes candidate.
-// - If later a "public" address appears and current is not public, replace it.
-// Returns:
-// - bestV4: best IPv4 "ip:port" if any
-// - bestV6: best IPv6 "[ip]:port" if any
-// - fallback: first valid address from list (stable)
-func PickBestV4V6FromLocalList(localStr string) (bestV4 string, bestV6 string, fallback string) {
-	addrs := SplitCommaAddrList(localStr)
-	if len(addrs) == 0 {
-		return "", "", ""
-	}
-	fallback = addrs[0]
-
-	var bestV4Public bool
-	var bestV6Public bool
-
-	for _, a := range addrs {
-		ip := ParseIPFromAddr(a)
-		if ip == nil {
-			continue
-		}
-		if ip.To4() != nil {
-			pub := IsPublicIPStrict(ip)
-			if bestV4 == "" {
-				bestV4 = a
-				bestV4Public = pub
-			} else if !bestV4Public && pub {
-				bestV4 = a
-				bestV4Public = true
-			}
-			continue
-		}
-		pub := IsPublicIPStrict(ip)
-		if bestV6 == "" {
-			bestV6 = a
-			bestV6Public = pub
-		} else if !bestV6Public && pub {
-			bestV6 = a
-			bestV6Public = true
-		}
-	}
-	return
-}
-
-// HasIPv6InLocalList reports whether localStr contains at least one valid IPv6 "[...]:port".
-func HasIPv6InLocalList(localStr string) bool {
-	_, v6, _ := PickBestV4V6FromLocalList(localStr)
-	return v6 != ""
-}
-
-// ChooseLocalAddrForPeer chooses ONE address from selfLocal to return to peer.
-// Rule (per your requirement):
-// - If both sides contain IPv6 -> return self IPv6 (best, public preferred)
-// - Else -> return self IPv4 (best, public preferred)
-// - Else -> return self IPv6 if any
-// - Else -> return fallback (first valid) or ""
-//
-// Works when:
-// - selfLocal is empty (returns "")
-// - selfLocal has 1 address (no comma)
-// - selfLocal has many addresses (comma list)
-// - peerLocal is empty (IPv6 mutual check fails, will choose IPv4 first)
-func ChooseLocalAddrForPeer(selfLocal, peerLocal string) string {
-	selfV4, selfV6, selfFallback := PickBestV4V6FromLocalList(selfLocal)
-	peerHasV6 := HasIPv6InLocalList(peerLocal)
-
-	if selfV6 != "" && peerHasV6 {
-		return selfV6
-	}
-	if selfV4 != "" {
-		return selfV4
-	}
-	if selfV6 != "" {
-		return selfV6
-	}
-	return selfFallback
-}
-
-// FixUdpListenAddrForRemote makes sure localAddr is a concrete address (not wildcard)
-// and matches the IP family of remoteAddr.
-// Returns (network, fixedLocalAddr).
-func FixUdpListenAddrForRemote(remoteAddr, localAddr string) (string, string, error) {
-	rip := ParseIPFromAddr(remoteAddr)
-	if rip == nil {
-		return "", "", fmt.Errorf("parse remote ip failed: %s", remoteAddr)
-	}
-	rip = NormalizeIP(rip)
-	wantV4 := rip.To4() != nil
-	network := "udp6"
-	if wantV4 {
-		network = "udp4"
-	}
-
-	port := GetPortStrByAddr(localAddr)
-	if port == "" || port == "0" {
-		return "", "", fmt.Errorf("invalid local port: %s", localAddr)
-	}
-
-	lip := ParseIPFromAddr(localAddr)
-	if lip != nil {
-		lip = NormalizeIP(lip)
-	}
-	localSpecified := lip != nil && !IsZeroIP(lip) && !lip.IsUnspecified()
-	localIsV4 := lip != nil && lip.To4() != nil
-
-	// already concrete and family matches
-	if localSpecified && (localIsV4 == wantV4) {
-		return network, localAddr, nil
-	}
-
-	var ip net.IP
-	var err error
-	if wantV4 {
-		ip, err = GetLocalUdp4IP()
-	} else {
-		ip, err = GetLocalUdp6IP()
-	}
-	if err != nil {
-		return "", "", err
-	}
-	ip = NormalizeIP(ip)
-	if ip == nil || IsZeroIP(ip) || ip.IsUnspecified() {
-		return "", "", fmt.Errorf("no usable local ip for %s", network)
-	}
-
-	return network, net.JoinHostPort(ip.String(), port), nil
-}
-
-func BuildTCPBindAddr(localIP string) net.Addr {
-	ip := net.ParseIP(strings.TrimSpace(localIP))
-	if ip == nil {
-		return nil
-	}
-	return &net.TCPAddr{IP: ip}
-}
-
-func BuildUDPBindAddr(localIP string) *net.UDPAddr {
-	ip := net.ParseIP(strings.TrimSpace(localIP))
-	if ip == nil {
-		return nil
-	}
-	return &net.UDPAddr{IP: ip}
-}
-
-func IsPublicHost(addr string) bool {
-	host := GetIpByAddr(addr)
-	if host == "" {
-		return false
-	}
-	if ip := net.ParseIP(host); ip != nil {
-		return IsPublicIPStrict(ip)
-	}
-	return true
 }
